robot/linux/requirements.txt
pyserial
pyrealsense2
pygame
opencv-python
openvino-dev[tensorflow2]
pytest
tflite-runtime